Separating implementation for ScalaZ and Cats.
