+++ util-validators-cats/src/main/scala/com/outworkers/util/validators/dsl/package.scala
@@ -1,0 +1,3 @@
+package com.outworkers.util.validators
+
+package object dsl extends Wrappers with ValidatorImplicits
